% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{datatable_codelist}
\alias{datatable_codelist}
\title{Helper function to create codelist / selected codes table}
\usage{
datatable_codelist(data, data_desc)
}
\description{
Helper function to create codelist / selected codes table
}
\keyword{internal}
