% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{get_codelist}
\alias{get_codelist}
\title{Get codelist from OpenCodelists}
\usage{
get_codelist(codelist_slug)
}
\arguments{
\item{codelist_slug}{String, specifying the slug of the codelist.
The naming convention of the codelist slug follows this structure: a \verb{<codelist-id>} is followed by / and a \verb{<version-id>}.
Note that the version ID is a sequence of 8 characters.
Some codelists may also have a version tag in the form of a date (YYYY-MM-DD) or a version number (e.g., v1.2) that can be used in place of the version ID.}
}
\description{
Get codelist from OpenCodelists
}
