% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{get_codelist_organisation}
\alias{get_codelist_organisation}
\title{Helper function to get the organisation for a codelist from OpenCodelists
This is important to use the API from OpenCodelists}
\usage{
get_codelist_organisation(codelist_slug)
}
\description{
Helper function to get the organisation for a codelist from OpenCodelists
This is important to use the API from OpenCodelists
}
\keyword{internal}
