% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{get_opencodelist}
\alias{get_opencodelist}
\title{Get codelist from OpenCodelists}
\usage{
get_opencodelist(codelist_url)
}
\arguments{
\item{codelist_url}{String, specifying URL to the codelist on OpenCodelists}
}
\description{
Get codelist from OpenCodelists
}
