% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opcs_usage}
\alias{opcs_usage}
\title{Yearly OPCS Code Usage from Hospital Admitted Patient Care Activity in England}
\format{
A data frame with 107,376 rows and 5 columns:
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{opcs_code}{The 4-character OPCS code.
Note that the punctuation from the code has been removed for compatibility with OpenCodelists.}
\item{usage}{Yearly summary of code usage.}
\item{description}{Description of OPCS Code}
}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity}
}
\usage{
opcs_usage
}
\description{
Yearly summary of OPCS code usage from 1st April 2013 to 31st March 2024.
All procedure codes are summarised to the 4-character level.
The code usage represents total annual count of each procedure, recorded across the primary and the secondary procedure positions.
Restricted codes for which annual usage is not published, have been removed.
}
\keyword{datasets}
