% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_individual}
\alias{plot_individual}
\title{Helper function to plot individual code usage}
\usage{
plot_individual(data)
}
\description{
Helper function to plot individual code usage
}
\keyword{internal}
