% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencodelists.R
\name{Codelist}
\alias{Codelist}
\title{Codelist}
\usage{
Codelist(
  .data = list(),
  row.names = NULL,
  coding_system = class_missing,
  full_slug = class_missing
)
}
\value{
New instance of class Codelist
}
\description{
Codelist
}
\section{Fields}{

\describe{
\item{\code{coding_system}}{String, specifying the coding system of the codelist.}

\item{\code{full_slug}}{Full slug of the codelist from OpenCodelists.}
}}

\keyword{internal}
