% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_usage}
\alias{icd10_usage}
\title{ICD-10 Code Usage from Hospital Admitted Patient Care Activity in England}
\format{
\subsection{\code{icd10_usage}}{

A data frame with 135,951 rows and 6 columns:
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{icd10_code}{ICD-10 COde.
Note that the punctuation from the code has been removed.}
\item{usage}{Yearly summary of code usage.}
\item{description}{Description of ICD-10 Code}
}
}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity}
}
\usage{
icd10_usage
}
\description{
Yearly summary of ICD-10 code usage from 1st April 2013 to 31st March 2024.
Restricted codes. for which annual usage is not published, have been removed from the dataset.
}
\keyword{datasets}
