% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opcs_usage}
\alias{opcs_usage}
\title{Yearly OPCS-4 Code Usage from Hospital Admitted Patient Care Activity in England}
\format{
A data frame with 107,376 rows and 5 columns:
\describe{
\item{start_date}{Start date of code usage count}
\item{end_date}{End date of code usage count}
\item{opcs_code}{The 4-character OPCS-4 code.
Note that the punctuation from the code has been removed for compatibility with OpenCodelists.}
\item{usage}{Annual count of code usage.}
\item{description}{Description of the OPCS-4 Code}
}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity}
}
\usage{
opcs_usage
}
\description{
Yearly summary of 4-character OPCS-4 code usage from 1st April 2013 to 31st March 2024.
The code usage represents the total annual count of each procedure, recorded across the primary and the secondary procedure positions.
Restricted codes for which annual usage is not published have been removed.
}
\examples{
# Filter to procedures involving "biopsy" after March 2020 (note each year runs April - March).
opcs_usage |>
dplyr:: filter(grepl("biopsy", description, ignore.case = TRUE) & lubridate:: year(end_date) > 2020)
}
\keyword{datasets}
